Update Rust 1.7->1.31 in CI
CI builds are failing because the `pkg-config` crate uses the `?` sigil,
which doesn't compile on older versions of Rust. Bumping the minimum
version to 1.31 should fix the build failures.

`libudev-sys` should still have no problem compiling on older versions
of Rust. To test that in CI would require locking dependencies to older
versions just for that build configuration.