Include Rust 1.4.0 in CI build matrix
The build matrix includes an old version of Rust to test compatibility
with older versions of the compiler. This is to ensure that the library
will be usable on old LTS releases that may have an older compiler.
However, this build configuration is fragile, because dependencies often
introduce the use of newer language constructs that are not available on
older versions of the compiler. This often results in bumping the
minimum Rust version in the build matrix, just to be compatible with
newer dependencies [1].

To make CI more stable with older compilers, a `Cargo.lock` is added to
lock the dependencies when running this build configuration. This works
with a general-purpose mechanism for symlinking files from
`ci/$TRAVIS_RUST_VERSION` before running a build.

[1]: 54bba8a427b25a2c62c6b4b259b9e15bd225911f